improve the params validation

+++ routes/items.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { itemModel, validateItem } = require('../models/item');
-const { OK, Created, Not_Found, Bad_Request } = require('./statusCode');
+const { OK, Created, Not_Found, Bad_Request, Not_Content } = require('./statusCode');
 
 const router = express.Router();
 
@@ -9,7 +9,7 @@ router.get('/', async (req, res) => {
     .find()
     .select('-__v')
     .sort('title');
-  res.send(items);
+  res.status(OK).send(items);
 });
 
 router.post('/', async (req, res) => {
@@ -26,6 +26,8 @@ router.put('/:id', async (req, res) => {
   const { error, value } = validateItem(req.body);
   if (error) return res.status(Bad_Request).json(error.details[0].message);
 
+  if (!req.params.id) return res.status(Bad_Request).json('The ID was not provided.');
+
   let item = await itemModel.findByIdAndUpdate(req.params.id, value, {
     new: true
   });
@@ -35,11 +37,13 @@ router.put('/:id', async (req, res) => {
 });
 
 router.delete('/:id', async (req, res) => {
+  if (!req.params.id) return res.status(Bad_Request).json('The ID was not provided.');
+
   let item = await itemModel.findByIdAndDelete(req.params.id);
 
   if (!item) return res.status(Not_Found).json('The item with the given ID was not found.');
 
-  res.status(OK).send(item);
+  res.status(Not_Content).end();
 });
 
 router.get('/:id', async (req, res) => {
